requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
array_record==0.8.1
astunparse==1.6.3
attrs==25.4.0
bddl==1.0.1
certifi==2025.10.5
charset-normalizer==3.4.4
chex==0.1.90
click==8.3.0
cloudpickle==2.1.0
contourpy==1.3.2
cycler==0.12.1
diffusers==0.35.2
dm-tree==0.1.9
docker-pycreds==0.4.0
docstring_parser==0.17.0
easydict==1.9
egl_probe==1.0.2
einops==0.4.1
etils==1.13.0
exceptiongroup==1.3.0
fastjsonschema==2.21.2
filelock==3.20.0
flatbuffers==25.9.23
flax==0.10.7
fonttools==4.60.1
fsspec==2025.9.0
ftfy==6.3.1
future==0.18.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
glfw==2.10.0
google-pasta==0.2.0
grpcio==1.76.0
gym==0.25.2
gym-notices==0.1.0
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
humanize==4.14.0
hydra-core==1.2.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
immutabledict==4.2.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
jax==0.6.2
jax-cuda12-pjrt==0.6.2
jax-cuda12-plugin==0.6.2
jaxlib==0.6.2
jaxtyping==0.3.3
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupytext==1.18.1
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
-e git+https://github.com/Lifelong-Robot-Learning/LIBERO.git@8f1084e3132a39270c3a13ebe37270a43ece2a01#egg=libero
llvmlite==0.45.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdit-py-plugins==0.5.0
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.2
mujoco==3.3.7
namex==0.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.62.1
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.14.0.64
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.28.7
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.6.0.66
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optax==0.2.6
optree==0.17.0
orbax-checkpoint==0.11.26
packaging==25.0
pathtools==0.1.2
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
promise==2.3
protobuf==6.33.0
psutil==7.1.2
pyarrow==22.0.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyOpenGL==3.1.10
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
robomimic==0.2.0
robosuite==1.4.0
rpds-py==0.28.0
safetensors==0.6.2
scipy==1.15.3
sentry-sdk==2.42.1
setproctitle==1.3.7
shortuuid==1.0.13
simple-parsing==0.1.7
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow-datasets==4.9.9
tensorflow-metadata==1.16.1
tensorflow_cpu==2.20.0
tensorstore==0.1.78
termcolor==3.2.0
thop==0.1.1.post2209072238
tokenizers==0.19.1
toml==0.10.2
tomli==2.3.0
toolz==1.1.0
torch==2.5.1+cpu
torchvision==0.20.1+cpu
tqdm==4.67.1
traitlets==5.14.3
transformers==4.40.2
treescope==0.1.10
triton==2.2.0
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uv==0.9.5
wadler_lindig==0.1.7
wandb==0.22.2
wcwidth==0.2.14
Werkzeug==3.1.3
wrapt==2.0.0
zipp==3.23.0
